% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_edits_batch.R
\name{create_multiEditR_report}
\alias{create_multiEditR_report}
\title{Creates an html report similar to the shiny app}
\usage{
create_multiEditR_report(
  batch_results,
  params,
  path = "./multiEditR_batch_results.html"
)
}
\arguments{
\item{batch_results}{The result of multiEditR::detect_edits_batch}

\item{params}{The parameters dataframe given to multiEditR::detect_edits_batch}

\item{path}{The file to write the html}
}
\description{
Creates an html report similar to the shiny app
}
