% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_edits_batch.R
\name{get_batch_stats_table}
\alias{get_batch_stats_table}
\title{get a single dataframe containing all test critical statistics}
\usage{
get_batch_stats_table(fits)
}
\arguments{
\item{fits}{The result of multiEditR::detect_edits_batch}
}
\value{
A data.frame
}
\description{
get a single dataframe containing all test critical statistics
}
