% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_control_chromatogram}
\alias{plot_control_chromatogram}
\title{Plot the chromatogram + percentage tiles for all detected motifs in the control}
\usage{
plot_control_chromatogram(fit)
}
\arguments{
\item{fit}{the result of detect_edits}
}
\value{
A ggplot object
}
\description{
Plot the chromatogram + percentage tiles for all detected motifs in the control
}
