% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_editing_barplot}
\alias{plot_editing_barplot}
\title{Plot statistics in the tested wt bases}
\usage{
plot_editing_barplot(fit)
}
\arguments{
\item{fit}{the result of detect_edits}
}
\value{
A ggplot object
}
\description{
Plot statistics in the tested wt bases
}
