% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_raw_sample}
\alias{plot_raw_sample}
\title{Plot the primary call percentage and trim / motif detection points}
\usage{
plot_raw_sample(fit)
}
\arguments{
\item{fit}{the result of detect_edits}
}
\value{
A ggplot object
}
\description{
Plot the primary call percentage and trim / motif detection points
}
