% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_sample_chromatogram}
\alias{plot_sample_chromatogram}
\title{Plot the chromatogram + percentage tiles for all detected motifs in the sample}
\usage{
plot_sample_chromatogram(fit)
}
\arguments{
\item{fit}{the result of detect_edits}
}
\value{
A ggplot object
}
\description{
Plot the chromatogram + percentage tiles for all detected motifs in the sample
}
