% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_trimmed_sample}
\alias{plot_trimmed_sample}
\title{Plot the primary call percentage just in the non-trimmed data}
\usage{
plot_trimmed_sample(fit)
}
\arguments{
\item{fit}{the result of detect_edits}
}
\value{
A ggplot object
}
\description{
Plot the primary call percentage just in the non-trimmed data
}
