% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{results}
\alias{results}
\title{Get complete fit results}
\usage{
results(fit)
}
\arguments{
\item{fit}{the result of detect_edits}
}
\value{
A dataframe with the following columns:
raw_sample_position - position in the sample sanger
passed_trimming - whether the sample sanger passed trimming (TRUE) or was low quality (FALSE)
motif_number - The numbered motif found in the control. -1 means not a match, 1 is the first match, 2 the second, etc.
edit_sig - whether the WT was significantly changed
control_base - The control sequence. Also serves as the motif sequence when motif_number > 0
sample_primary_call - The highest defined peak in the sample
sample_secondary_call - The second highest defined peak in the sample
A-T_area - The maximum sample height of the base peak
A-T_perc - The percent of this base peak's height over all bases
edit_pvalue/adjust 
motif_seq - the motif_seq as used. Meaning, if motif_fwd == FALSE, the rev-com of the motif you supplied
sample_file
control_file
}
\description{
Get complete fit results
}
