% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_edits_batch.R
\name{save_batch_skeleton}
\alias{save_batch_skeleton}
\title{Saves an example parameters excel spreadsheet at the given location}
\usage{
save_batch_skeleton(path)
}
\arguments{
\item{path}{A writable file path, e.g. params.xlsx}
}
\description{
Saves an example parameters excel spreadsheet at the given location
}
