% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_edits.R
\name{detect_edits}
\alias{detect_edits}
\title{the primary function to run edit detection.}
\usage{
detect_edits(
  sample_file,
  ctrl_file,
  motif,
  motif_fwd,
  wt,
  edit,
  phred_cutoff = 1e-05,
  p_value = 0.01
)
}
\arguments{
\item{sample_file}{path to the sample .ab1 file}

\item{ctrl_file}{path to the control .ab1 or .fa(sta) file}

\item{motif}{string holding the motif of interest}

\item{motif_fwd}{boolean, true if the motif should be in the control as-is. False if the motif requires rev-com.}

\item{wt}{the base to be tested for editing in motif locations}

\item{edit}{the predicted edit for the base}

\item{phred_cutoff}{the cutoff for trimming. lower is more stringent. default = 0.001}

\item{p_value}{the cutoff for BH-adjusted significance.}
}
\value{
A multiEditR object (list of lists)
}
\description{
the primary function to run edit detection.
}
\examples{

sample_file = system.file("extdata", "RP272_cdna_wt.ab1", package="multiEditR")
ctrl_file = system.file("extdata", "RP272_cdna_ko.ab1", package="multiEditR")
motif = "AGTAGCTGGGATTACAGATG"
fit = detect_edits(sample_file, ctrl_file,
            motif = motif, motif_fwd = TRUE,
            wt = "A", edit = "G",
            phred_cutoff = 0.001, p_value = 0.05)
            
tbl = results(fit)
writexl::write_xlsx(tbl, "my_results.xlsx") 

plot_sample_chromatogram(fit)
plot_raw_sample(fit)       
            

}
